Use "Double" instead of "Float" for "float" in Kotlin

+++ spaghetti-kotlin-support/src/main/groovy/com/prezi/spaghetti/kotlin/AbstractKotlinGeneratorVisitor.groovy
@@ -22,7 +22,7 @@ abstract class AbstractKotlinGeneratorVisitor extends StringModuleVisitorBase {
 	protected static final EnumMap<PrimitiveType, String> PRIMITIVE_TYPES = [
 			(PrimitiveType.BOOL): "Bool",
 			(PrimitiveType.INT): "Int",
-			(PrimitiveType.FLOAT): "Float",
+			(PrimitiveType.FLOAT): "Double",
 			(PrimitiveType.STRING): "String",
 			(PrimitiveType.ANY): "Any"
 	]
